Move clear_save to test_helper.py
